+++ src/main/java-interfaces/example_interfaces/msg/dds/MultiArrayLayout.java
@@ -71,7 +71,7 @@ public class MultiArrayLayout implements ROS2Message<MultiArrayLayout>
    public static final java.lang.String name = "example_interfaces::msg::dds_::MultiArrayLayout_";
 
    private final IDLObjectSequence<example_interfaces.msg.dds.MultiArrayDimension> dim_; // Array of dimension properties
-   private long data_offset_; // padding bytes at front of data
+   private int data_offset_; // padding bytes at front of data
 
    public MultiArrayLayout()
    {
@@ -94,7 +94,7 @@ public class MultiArrayLayout implements ROS2Message<MultiArrayLayout>
    public void serialize(CDRBuffer buffer)
    {
       dim_.serialize(buffer);
-      buffer.writeLong(data_offset_);
+      buffer.writeInt(data_offset_);
 
    }
 
@@ -102,7 +102,7 @@ public class MultiArrayLayout implements ROS2Message<MultiArrayLayout>
    public void deserialize(CDRBuffer buffer)
    {
       dim_.deserialize(buffer);
-      data_offset_ = buffer.readLong();
+      data_offset_ = buffer.readInt();
 
    }
 
@@ -119,12 +119,12 @@ public class MultiArrayLayout implements ROS2Message<MultiArrayLayout>
       return dim_;
    }
 
-   public long getDataOffset()
+   public int getDataOffset()
    {
       return data_offset_;
    }
 
-   public void setDataOffset(long data_offset_)
+   public void setDataOffset(int data_offset_)
    {
       this.data_offset_ = data_offset_;
    }
