+++ src/main/java-interfaces/example_interfaces/msg/dds/UInt16.java
@@ -28,7 +28,7 @@ public class UInt16 implements ROS2Message<UInt16>
       it into a system you should create a semantically meaningful message type.
       If you want to embed it in another message, use the primitive data type instead.
    */
-   private int data_;
+   private short data_;
 
    public UInt16()
    {
@@ -47,14 +47,14 @@ public class UInt16 implements ROS2Message<UInt16>
    @Override
    public void serialize(CDRBuffer buffer)
    {
-      buffer.writeInt(data_);
+      buffer.writeShort(data_);
 
    }
 
    @Override
    public void deserialize(CDRBuffer buffer)
    {
-      data_ = buffer.readInt();
+      data_ = buffer.readShort();
 
    }
 
@@ -65,12 +65,12 @@ public class UInt16 implements ROS2Message<UInt16>
 
    }
 
-   public int getData()
+   public short getData()
    {
       return data_;
    }
 
-   public void setData(int data_)
+   public void setData(short data_)
    {
       this.data_ = data_;
    }
