+++ src/main/java-interfaces/example_interfaces/msg/dds/UInt32.java
@@ -28,7 +28,7 @@ public class UInt32 implements ROS2Message<UInt32>
       it into a system you should create a semantically meaningful message type.
       If you want to embed it in another message, use the primitive data type instead.
    */
-   private long data_;
+   private int data_;
 
    public UInt32()
    {
@@ -47,14 +47,14 @@ public class UInt32 implements ROS2Message<UInt32>
    @Override
    public void serialize(CDRBuffer buffer)
    {
-      buffer.writeLong(data_);
+      buffer.writeInt(data_);
 
    }
 
    @Override
    public void deserialize(CDRBuffer buffer)
    {
-      data_ = buffer.readLong();
+      data_ = buffer.readInt();
 
    }
 
@@ -65,12 +65,12 @@ public class UInt32 implements ROS2Message<UInt32>
 
    }
 
-   public long getData()
+   public int getData()
    {
       return data_;
    }
 
-   public void setData(long data_)
+   public void setData(int data_)
    {
       this.data_ = data_;
    }
