+++ src/main/java-interfaces/example_interfaces/msg/dds/UInt8.java
@@ -28,7 +28,7 @@ public class UInt8 implements ROS2Message<UInt8>
       it into a system you should create a semantically meaningful message type.
       If you want to embed it in another message, use the primitive data type instead.
    */
-   private short data_;
+   private byte data_;
 
    public UInt8()
    {
@@ -47,14 +47,14 @@ public class UInt8 implements ROS2Message<UInt8>
    @Override
    public void serialize(CDRBuffer buffer)
    {
-      buffer.writeShort(data_);
+      buffer.writeByte(data_);
 
    }
 
    @Override
    public void deserialize(CDRBuffer buffer)
    {
-      data_ = buffer.readShort();
+      data_ = buffer.readByte();
 
    }
 
@@ -65,12 +65,12 @@ public class UInt8 implements ROS2Message<UInt8>
 
    }
 
-   public short getData()
+   public byte getData()
    {
       return data_;
    }
 
-   public void setData(short data_)
+   public void setData(byte data_)
    {
       this.data_ = data_;
    }
